requirements.txt
Flask
--find-links https://download.pytorch.org/whl/cpu
torch == 2.0.1
torchvision == 0.15.2
opencv-python ~= 4.9.0
pillow
gdown ~= 5.2.0
python-dotenv ~= 1.0.0
ultralytics
transformers
supervision
pycocotools
yapf
addict
timm
onnx == 1.13.0